comms/requirements.txt
cryptography
fastapi
pydantic
PyJWT
uvicorn
